Fix potential vulnerability in the rand.Seed
According to webauthn spec. the challange needs to be hard to guess
otherwise the protocol is vulnerable to reply attack see:
https://www.w3.org/TR/webauthn-2/#sctn-cryptographic-challenges

> In order to prevent replay attacks, the challenges MUST contain
> enough entropy to make guessing them infeasible. Challenges
> SHOULD therefore be at least 16 bytes long.

Initating with the random seed with time makes it easy to guess after
a restart. The space of seeds can be quickly enumerated, because time
has lager resultion than nanosecods. See this more details:
https://github.com/Quiq/webauthn_proxy